Commit 2

+++ codigos/borarchi.php
@@ -1,22 +1,18 @@
 <?php
-	$archivo = $_GET['archi'];
-	session_start();
-
-	$ruta = "d:\\mybox";
-    $ruta = $ruta.'/'.$_SESSION["usuario"].'/'.$archivo;
-
-	/*Se intenta eliminar un fichero y se informa del resultado.*/
-    echo "<h3>";
-		if (@unlink($ruta)){
-			echo ("Se ha eliminado el fichero.");
-		} else {
-			echo ("NO se pudo eliminar el fichero.");
-		}
-	echo "</h3>";
-
-	//Retorna al punto de invocación
-	$Ir_A = $_SERVER["HTTP_REFERER"];
-	echo "<script language='JavaScript'>";
-	echo "location.href='".$Ir_A."'";
-	echo "</script>";
+$archivo = $_GET['archi'];
+session_start();
+$ruta = "C:\\myboxusers\\" . $_SESSION["usuario"] . '\\' . $archivo;
+echo "<h3>";
+if (file_exists($ruta)) {
+    if (@unlink($ruta)) {
+        echo "Se ha eliminado el fichero.";
+    } else {
+        echo "NO se pudo eliminar el fichero.";
+    }
+} else {
+    echo "El fichero no existe.";
+}
+echo "</h3>";
+header("Location: /mybox/carpetas.php" . (isset($_GET['folder']) ? '?folder=' . $_GET['folder'] : ''));
+exit();
 ?>
